requirements.txt
autopep8==1.5.5
cycler==0.10.0
kiwisolver==1.3.1
llvmlite==0.35.0
matplotlib==3.3.4
numba==0.52.0
numpy==1.20.1
Pillow==8.1.2
pycodestyle==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
toml==0.10.2
